refactor: Move AccountsTree component to tabs directory for better organization
