+++ src/components/tabs/AccountsTree.tsx
@@ -1,20 +1,20 @@
 import { CircleOff, Delete, ListTree, Loader2, PenLine, PlusCircle, Unlink, } from "lucide-react";
 import { useEffect, useMemo, useState } from "react";
-import { Button } from "./ui/button";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "./ui/card";
-import { Label } from "./ui/label";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
+import { Label } from "@/components/ui/label";
 import { TypedApi, SS58String, Binary } from "polkadot-api";
 import { ChainId } from "@reactive-dot/core";
 import { ChainDescriptorOf } from "@reactive-dot/core/internal.js";
 import { LoadingPlaceholder } from "~/pages/Loading";
 import { AccountTreeNode } from "~/hooks/UseAccountsTree";
 import { DialogMode, OpenTxDialogArgs_modeSet } from "~/types";
-import { Badge } from "./ui/badge";
+import { Badge } from "@/components/ui/badge";
 import { fetchSuperOf, prepareRawSetSubs } from "~/utils/subaccounts";
-import { AccountSelector } from "./ui/account-selector";
-import { Input } from "./ui/input";
+import { AccountSelector } from "@/components/ui/account-selector";
+import { Input } from "@/components/ui/input";
 import { Identity } from "~/types/Identity";
-import { Alert, AlertTitle, AlertDescription } from "./ui/alert";
+import { Alert, AlertTitle, AlertDescription } from "@/components/ui/alert";
 import { useFormatAmount } from "~/hooks/useFormatAmount";
 import { ChainInfo } from "~/store/ChainStore";
 import { useWalletAccounts } from "~/hooks/useWalletAccounts";
